Rename all headers for name.h
