+++ include/ics3/check_invalid.h
@@ -24,8 +24,8 @@ CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-#ifndef LIBICS3_ICS3_CHECK_INVALID_H_
-#define LIBICS3_ICS3_CHECK_INVALID_H_
+#ifndef LIBICS3_ICS3_CHECK_INVALID_H
+#define LIBICS3_ICS3_CHECK_INVALID_H
 
 #include <stdexcept>
 #include <utility>
@@ -49,4 +49,4 @@ constexpr T checkValidRange(const T input, const T min, const T max)
 }
 }
 
-#endif // LIBICS3_ICS3_CHECK_INVALID_H_
+#endif // LIBICS3_ICS3_CHECK_INVALID_H
