+++ include/ics3/ics
@@ -24,16 +24,16 @@ CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-#ifndef LIBICS3_ICS3_ICS_H_
-#define LIBICS3_ICS3_ICS_H_
+#ifndef LIBICS3_ICS3_ICS_H
+#define LIBICS3_ICS3_ICS_H
 
-#include "ics3/angle.hpp"
-#include "ics3/baudrate.hpp"
-#include "ics3/check_invalid.hpp"
-#include "ics3/eepparam.hpp"
+#include "ics3/angle.h"
+#include "ics3/baudrate.h"
+#include "ics3/check_invalid.h"
+#include "ics3/eepparam.h"
 #include "ics3/eeprom.h"
 #include "ics3/ics3.h"
-#include "ics3/id.hpp"
-#include "ics3/parameter.hpp"
+#include "ics3/id.h"
+#include "ics3/parameter.h"
 
-#endif // LIBICS3_ICS3_ICS_H_
+#endif // LIBICS3_ICS3_ICS_H
